+++ social_core/pipeline/user.py
@@ -6,8 +6,8 @@ from ..utils import slugify, module_member
 USER_FIELDS = ['username', 'email']
 
 
-def get_username(strategy, details, user=None, *args, **kwargs):
-    if 'username' not in strategy.setting('USER_FIELDS', USER_FIELDS):
+def get_username(strategy, details, backend, user=None, *args, **kwargs):
+    if 'username' not in backend.setting('USER_FIELDS', USER_FIELDS):
         return
     storage = strategy.storage
 
@@ -61,12 +61,12 @@ def get_username(strategy, details, user=None, *args, **kwargs):
     return {'username': final_username}
 
 
-def create_user(strategy, details, user=None, *args, **kwargs):
+def create_user(strategy, details, backend, user=None, *args, **kwargs):
     if user:
         return {'is_new': False}
 
     fields = dict((name, kwargs.get(name, details.get(name)))
-                  for name in strategy.setting('USER_FIELDS', USER_FIELDS))
+                  for name in backend.setting('USER_FIELDS', USER_FIELDS))
     if not fields:
         return